- Changed: GameObject.Realm has now changed from a byte to an eRealm (which can still be cast to a byte).
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk/GameServer@954 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ commands/gmcommands/keepguard.cs
@@ -149,7 +149,7 @@ namespace DOL.GS.Commands
 							guard.Y = client.Player.Y;
 							guard.Z = client.Player.Z;
 							guard.Heading = client.Player.Heading;
-							guard.Realm = (byte)guard.CurrentZone.GetRealm();
+							guard.Realm = guard.CurrentZone.GetRealm();
 							guard.SaveIntoDatabase();
 							foreach (AbstractArea area in guard.CurrentAreas)
 							{
